refactor callback proxy and registry
