+++ evm/src/main/java/io/horizen/evm/LibEvm.java
@@ -1,17 +1,20 @@
 package io.horizen.evm;
 
 import com.fasterxml.jackson.databind.type.TypeFactory;
+import com.sun.jna.Callback;
 import com.sun.jna.Native;
 import com.sun.jna.Pointer;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.nio.charset.StandardCharsets;
+
 final class LibEvm {
-    static native void FreeBuffer(Pointer ptr);
+    private static native void FreeBuffer(Pointer ptr);
 
-    static native Pointer CreateBuffer(int size);
+    private static native Pointer CreateBuffer(int size);
 
-    private static native void SetCallbackProxy(LibEvmCallback.CallbackProxy callback);
+    private static native void SetCallbackProxy(CallbackProxy callback);
 
     private static native void SetupLogging(int callbackHandle, String level);
 
@@ -19,6 +22,11 @@ final class LibEvm {
 
     private static final Logger logger = LogManager.getLogger();
     private static final GlogCallback logCallback = new GlogCallback(logger);
+
+    // this singleton instance of the callback will be passed to libevm,
+    // the static reference here will also prevent the callback instance from being garbage collected,
+    // because without it the only reference might be from native code (libevm) and the GC does not know about that
+    private static final CallbackProxy proxy = new CallbackProxy();
 
     private static String getOSLibExtension() {
         var os = System.getProperty("os.name").toLowerCase();
@@ -37,7 +45,7 @@ final class LibEvm {
         // bind native methods in this class to libevm
         Native.register(libName);
         // register callback
-        SetCallbackProxy(LibEvmCallback.proxy);
+        SetCallbackProxy(proxy);
         // propagate log4j log level to glog
         SetupLogging(logCallback.handle, GlogCallback.log4jToGlogLevel(logger.getLevel()));
     }
@@ -113,5 +121,30 @@ final class LibEvm {
      */
     static void invoke(String method) {
         invoke(method, null, Void.class);
+    }
+
+    private static class CallbackProxy implements Callback {
+        public Pointer callback(int handle, Pointer msg) {
+            try {
+                // we do not need to free the Pointer here, as it is freed on the libevm side when the callback returns
+                var result = CallbackRegistry.invoke(handle, msg.getString(0));
+                if (result == null) return null;
+                var bytes = result.getBytes(StandardCharsets.UTF_8);
+                // allocate buffer on native side and write the string into it,
+                // length plus one because the string needs to be null-terminated
+                var ptr = LibEvm.CreateBuffer(bytes.length + 1);
+                ptr.write(0, bytes, 0, bytes.length);
+                // make absolutely sure the string is null-terminated,
+                // the buffer is zero-initialized so this should be redundant
+                ptr.setByte(bytes.length, (byte) 0);
+                // note: this buffer is expected to be freed on the native side
+                return ptr;
+            } catch (Exception e) {
+                // note: make sure we do not throw any exception here because this callback is called by native code
+                // for diagnostics we log the exception here
+                logger.warn("error while handling callback from libevm", e);
+            }
+            return null;
+        }
     }
 }
